refactor(web): extract SolveOverlays from DesignSession
Part of issue #642 seam 5b-2. Pulls the busy bar/cancel, backend-
disallowed banner, poor-match banner, and solve-error banner into a
module-level SolveOverlays component. Per the design comment, the
backend-switch button's `backendTouchedRef.current = true;
setSlotBackend(...)` pair folds into an onSwitchBackend callback that
DesignSession passes in — the ref write stays a DesignSession concern,
not something the presentational child touches directly.
